aya: move Mmap struct to aya::util
